0.1.5

+++ pylabel/exporter.py
@@ -223,7 +223,7 @@ class Export():
         output_file_paths = []
 
         yolo_dataset = ds.df.copy(deep=True)
-        yolo_dataset.cat_id = yolo_dataset.cat_id.astype("Int64")
+        yolo_dataset.cat_id = yolo_dataset.cat_id.astype('float').astype('Int32')
         yolo_dataset["center_x_scaled"] = (yolo_dataset["ann_bbox_xmin"] + (yolo_dataset["ann_bbox_width"]*0.5))/yolo_dataset["img_width"]
         yolo_dataset["center_y_scaled"] = (yolo_dataset["ann_bbox_ymin"] + (yolo_dataset["ann_bbox_height"]*0.5))/yolo_dataset["img_height"]
         yolo_dataset["width_scaled"] = yolo_dataset["ann_bbox_width"] / yolo_dataset["img_width"]
